# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ilm School Portal API
  version: 1.0.0
  description: |-
    Multi-school (multi-tenant) school management API. Every role except
    `system_admin` is scoped to exactly one school (`schoolId` on the JWT);
    `system_admin` is a platform-level role with no school of its own and
    can act across every school.

    **Auth**: session is an httpOnly `token` cookie (JWT), set by
    `POST /auth/login` and cleared by `POST /auth/logout`. There is no
    bearer-token flow — the cookie is sent automatically by the browser on
    same-origin requests (`withCredentials: true` on the frontend's axios
    client). A request with no cookie, an invalid/expired token, or a
    revoked token (see below) gets `401`.

    **Session lifetime**: tokens are signed for 1 hour. Any authenticated
    request past the halfway point of that window silently reissues the
    cookie, so an active user is never hard-logged-out mid-session — only
    genuine inactivity lets it expire. That sliding renewal is itself capped
    at 12 hours from the original login (`sessionStartedAt` in the token),
    so a session can't be kept alive indefinitely just by staying active —
    after 12 hours the user has to log in again regardless of activity.
    Logging out invalidates the token server-side immediately (via a
    `token_version` check on every request), not just client-side — and
    because there's no per-device session tracking, logging out on one
...
    sessions.

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /teacher/attendance/report:
    get:
      tags:
        - Teacher
      summary: Download an attendance report (CSV) for one of the teacher's own classes
      description: Gated by `attendance_report`. Always the last 365 days — no custom range.
      parameters:
        - name: classId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: classId is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Feature disabled, or teacher not assigned to this class'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getTeacherAttendanceReport
  '/teacher/attendance/{classId}':
    get:
      tags:
        - Teacher
      summary: Get the attendance register for a class on a given date
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Defaults to today
      responses:
        '200':
          description: Register
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  students:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        first_name:
                          type: string
                        surname:
                          type: string
                        attendance_status:
                          type: string
                          enum:
                            - PRESENT
                            - ABSENT
                          nullable: true
                          description: null if not yet marked for this date
        '400':
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Teacher not assigned to this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getTeacherAttendanceByClassId
  '/teacher/attendance/{classId}/history':
    get:
      tags:
        - Teacher
      summary: Get daily attendance summary counts for a class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: History
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        present_count:
                          type: integer
                        absent_count:
                          type: integer
                        total_count:
                          type: integer
        '403':
          description: Teacher not assigned to this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: getTeacherAttendanceByClassIdHistory
  /teacher/attendance/mark:
    post:
      tags:
        - Teacher
      summary: Mark one student's attendance for a class/date
      description: Upsert — repeated calls for the same class/student/date overwrite the status.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - classId
                - studentId
                - status
              properties:
                classId:
                  type: integer
                studentId:
                  type: integer
                status:
                  type: string
                  enum:
                    - present
                    - absent
                    - PRESENT
                    - ABSENT
                  description: Case-insensitive
                date:
                  type: string
                  format: date
                  description: Defaults to today
      responses:
        '200':
          description: Marked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  date:
                    type: string
                    format: date
        '400':
          description: 'Missing fields, or invalid status/date'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Teacher not assigned to this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postTeacherAttendanceMark
  /teacher/tasks/create:
    post:
      tags:
        - Teacher
      summary: 'Create a task for a class, or independent homework for one student'
      description: |-
        Pass `studentId` to target one student instead of the whole class
        ("independent homework") — this is gated by the `student_notes`
        flag since it lives on the same UI panel as student notes.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - classId
                - title
              properties:
                classId:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                due_date:
                  type: string
                  format: date
                studentId:
                  type: integer
                  description: Optional — targets one student in the class instead of the whole class
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '400':
          description: studentId given but that student isn't in this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Teacher not assigned to this class, or independent homework disabled'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: postTeacherTasksCreate
  '/teacher/tasks/{id}':
    delete:
      tags:
        - Teacher
      summary: Delete a task
      description: Only permitted for a task belonging to one of the teacher's own classes.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/MessageOk'
        '403':
          description: Teacher not assigned to this task's class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: deleteTeacherTasksById
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly JWT set by POST /auth/login. See the top-level description for session lifetime/refresh/revocation behavior.
...
  responses:
    MessageOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
...